refactor: MenuItem을 MenuWideItem으로 common에 추가
