+++ fourtoon-cookie/src/components/common/MenuWideButton/MenuWideButton.styled.ts
